Add routes items and newItem
This commit adds routes for the main components of
our project: items and newItem.

Users can click the logo of the app to return to
the homepage, click "Items" to go to the items
page (which is currently the homepage) and "New Items"
to go to that page.

No CSS has been added and needs to be added in
future commits.